Remove cycles
