+++ package.json
@@ -27,7 +27,7 @@
   "files": ["dist"],
   "devDependencies": {
     "@alinea/suite": "^0.6.2",
-    "@biomejs/biome": "^1.8.3",
+    "@biomejs/biome": "^1.9.4",
     "@cloudflare/workers-types": "^4.20230628.0",
     "@electric-sql/pglite": "^0.2.12",
     "@libsql/client": "^0.10.0",
