+++ src/core/expr/Input.ts
@@ -9,7 +9,6 @@ import {
   hasTable
 } from '../Internal.ts'
 import {type Sql, sql} from '../Sql.ts'
-import {mapToColumn} from '../query/Shared.ts'
 
 export type Input<T = unknown> = HasSql<T> | T
 
@@ -26,4 +25,15 @@ export function input<T>(value: Input<T>, maybeField?: Input<T>): Sql<T> {
   return fieldSource && 'mapToDriverValue' in fieldSource
     ? mapToColumn(fieldSource as ColumnData, value)
     : sql.value(value)
+}
+
+export function mapToColumn<T>(
+  {mapToDriverValue}: ColumnData,
+  expr: Input<T>
+): Sql<T> {
+  const isObject = expr && typeof expr === 'object'
+  if (isObject && hasSql(expr)) return getSql(expr)
+  return input(
+    expr !== null && mapToDriverValue ? mapToDriverValue(expr) : expr
+  ) as Sql<T>
 }
